// ProcurementToPay-FE-main/src/app/core/models/role.ts
export enum Role {
  Admin = "ROLE_ADMIN",
  User = "ROLE_USER",
  Financial = "ROLE_FINANCIAL",
  Technical = "ROLE_TECHNICAL",
  Desk = "ROLE_DESK",
  HOD = "ROLE_HEAD",
  Supplier = "ROLE_SUPPLIER",
  ProcAdmin = "ROLE_PROCUREMENT_ADMIN",
}

// Clerk = "ROLE_CLERK",
  // Executive = "ROLE_EXECUTIVE",
  // Supervisor = "ROLE_SUPERVISOR"